requirements.txt
app~=0.0.1
boto3~=1.35.80
botocore~=1.35.80
fastapi~=0.115.6
ipdb~=0.13.13
Jinja2~=3.1.4
matplotlib~=3.9.4
numpy~=2.2.0
openai~=1.57.4
pandas~=2.2.3
passlib~=1.7.4
pip~=23.3
psycopg2-binary
pydantic~=2.10.3
pydantic[email]
pydantic_settings
PyJWT~=2.10.1
python-jose~=3.3.0
python-dotenv~=1.0.1
reportlab~=4.2.5
replit~=4.1.0
seaborn~=0.13.2
SQLAlchemy~=2.0.32
typing_extensions~=4.12.2
uvicorn~=0.32.1